refactor(app): extract scrollbar math to own module (phase 1i)
Two visual-row-aware scrollbar helpers in input.rs (calculate_scrollbar_
jump_visual, calculate_scrollbar_drag_relative_visual, ~170 lines total)
were already written as associated functions: no `&self`, called via
`Self::calculate_scrollbar_*(&mut state.buffer, ...)`. Pure math dressed
up as methods.

Move them to a new app/scrollbar_math module as free functions
(scrollbar_jump_visual, scrollbar_drag_relative_visual). Collapse the
large duplicated "build visual row map" block into a shared
build_visual_row_map helper and a small position_at lookup, which
together dedupe ~30 lines between the two functions. Both callers in
input.rs switch from Self::calculate_scrollbar_*(...) to
super::scrollbar_math::scrollbar_*(...).

The &mut Buffer parameter remains because Buffer::line_iterator needs
mutable access for lazy-load internals; nothing else touches Editor.

Part of phase 1 (free pure helpers) of
docs/internal/editor-modules-refactor-plan.md. With this, phase 1 is
complete: ~1500 lines of pure logic are out of `impl Editor`, in 8
subsystem-agnostic modules, with unit tests that build nothing more
than plain data structures.
